sales-agent/requirements.txt

google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
langchain==0.0.245
langsmith==0.0.15
lxml==4.9.3
openai>=0.27.8
openapi-schema-pydantic>=1.2.4
tqdm>=4.65.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
pytest>=7.4.0
chromadb>=0.4.3
tiktoken>=0.4.0
pytest-asyncio